Replace common properties to XTFConfiguration, delete RadImages this property will be used in radanalytics-utils;
